Fix Encryption Keys
