+++ UnityProject/Assets/Scripts/Items/EncryptionKey.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using UnityEngine;
 using Mirror;
 using Objects.Telecomms;
+using Messages.Client;
 using ScriptableObjects.Communications;
 
 namespace Items
@@ -12,7 +13,7 @@ namespace Items
 /// <summary>
 ///     Encryption Key properties
 /// </summary>
-public class EncryptionKey : NetworkBehaviour, ICheckedInteractable<HandApply>, ICheckedInteractable<MouseDrop>
+public class EncryptionKey : NetworkBehaviour, ICheckedInteractable<HandApply>, ICheckedInteractable<MouseDrop>, IClientInteractable<InventoryApply>
 {
 	//TODO (Max): Turn this into a list and support multiple encryption data in one key
 	public EncryptionDataSO EncryptionDataSo;
@@ -134,33 +135,6 @@ public class EncryptionKey : NetworkBehaviour, ICheckedInteractable<HandApply>, 
 		return channelsByMask;
 	}
 
-	public Sprite binarySprite;
-	public Sprite captainSprite;
-	public Sprite centCommSprite;
-	public Sprite chiefEngineerSprite;
-	public Sprite chiefMedicalOfficerSprite;
-
-	public Sprite commonSprite;
-	public Sprite engineeringSprite;
-	public Sprite headOfPersonnelSprite;
-	public Sprite headOfSecuritySprite;
-	public Sprite medicalSprite;
-	public Sprite quarterMasterSprite;
-	public Sprite researchDirectorSprite;
-	public Sprite scienceSprite;
-	public Sprite securitySprite;
-
-	public Sprite serviceSprite;
-
-	//So that we don't have to create a different item for each type
-	public SpriteRenderer spriteRenderer;
-
-	public Sprite supplySprite;
-	public Sprite syndicateSprite;
-	public Sprite srvsecSprite;
-	public Sprite centCommPlusSprite;
-	public Sprite srvmedSprite;
-
 	[SerializeField] //to show in inspector
 	private EncryptionKeyType type;
 
@@ -175,13 +149,13 @@ public class EncryptionKey : NetworkBehaviour, ICheckedInteractable<HandApply>, 
 				Logger.LogError("Encryption keys cannot be None type!", Category.Chat);
 				type = EncryptionKeyType.Common;
 			}
-			UpdateSprite();
+			//UpdateSprite();
 		}
 	}
 
 	private void Start()
 	{
-		UpdateSprite();
+
 	}
 
 /// Look ma, no syncvars!
@@ -200,80 +174,6 @@ public class EncryptionKey : NetworkBehaviour, ICheckedInteractable<HandApply>, 
 		type = keyType;
 		base.OnDeserialize(reader, initialState);
 	}
-
-	#region Set the sprite based on key type
-
-	private void UpdateSprite()
-	{
-		switch (type)
-		{
-			case EncryptionKeyType.Binary:
-				spriteRenderer.sprite = binarySprite;
-				break;
-			case EncryptionKeyType.Captain:
-				spriteRenderer.sprite = captainSprite;
-				break;
-			case EncryptionKeyType.Supply:
-				spriteRenderer.sprite = supplySprite;
-				break;
-			case EncryptionKeyType.CentComm:
-				spriteRenderer.sprite = centCommSprite;
-				break;
-			case EncryptionKeyType.ChiefEngineer:
-				spriteRenderer.sprite = chiefEngineerSprite;
-				break;
-			case EncryptionKeyType.ChiefMedicalOfficer:
-				spriteRenderer.sprite = chiefMedicalOfficerSprite;
-				break;
-			case EncryptionKeyType.Common:
-				spriteRenderer.sprite = commonSprite;
-				break;
-			case EncryptionKeyType.Engineering:
-				spriteRenderer.sprite = engineeringSprite;
-				break;
-			case EncryptionKeyType.HeadOfPersonnel:
-				spriteRenderer.sprite = headOfPersonnelSprite;
-				break;
-			case EncryptionKeyType.HeadOfSecurity:
-				spriteRenderer.sprite = headOfSecuritySprite;
-				break;
-			case EncryptionKeyType.Medical:
-				spriteRenderer.sprite = medicalSprite;
-				break;
-			case EncryptionKeyType.QuarterMaster:
-				spriteRenderer.sprite = quarterMasterSprite;
-				break;
-			case EncryptionKeyType.ResearchDirector:
-				spriteRenderer.sprite = researchDirectorSprite;
-				break;
-			case EncryptionKeyType.Science:
-				spriteRenderer.sprite = scienceSprite;
-				break;
-			case EncryptionKeyType.Security:
-				spriteRenderer.sprite = securitySprite;
-				break;
-			case EncryptionKeyType.Service:
-				spriteRenderer.sprite = serviceSprite;
-				break;
-			case EncryptionKeyType.Syndicate:
-				spriteRenderer.sprite = syndicateSprite;
-				break;
-			case EncryptionKeyType.SrvSec:
-				spriteRenderer.sprite = srvsecSprite;
-				break;
-			case EncryptionKeyType.SrvMed:
-				spriteRenderer.sprite = srvmedSprite;
-				break;
-			case EncryptionKeyType.CentCommPlus:
-				spriteRenderer.sprite = centCommPlusSprite;
-				break;
-			default:
-				spriteRenderer.sprite = commonSprite;
-				break;
-		}
-	}
-
-	#endregion
 
 	public void onExamine(Vector3 worldPos)
 	{
@@ -308,6 +208,19 @@ public class EncryptionKey : NetworkBehaviour, ICheckedInteractable<HandApply>, 
 		{
 			radio.AddEncryptionKey(this);
 		}
+	}
+
+	public bool Interact(InventoryApply interaction)
+	{
+		//insert the headset key if this is used on a headset
+		if (interaction.UsedObject == gameObject
+		    && interaction.TargetObject.GetComponent<Headset>() != null)
+		{
+			UpdateHeadsetKeyMessage.Send(interaction.TargetObject, gameObject);
+			return true;
+		}
+
+		return false;
 	}
 }
 }
