--- UnityProject/Assets/Scripts/Items/HeadsetKey.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5cf506c5a1f18344a80b82acaac00e23
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
